backend/pyproject.toml
[project]
name = "vostkrep"
version = "0.1.0"
description = ""
authors = [
    {name = "Ксения",email = "kssshaab@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "django (>=4.2,<6.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "poetry-core (>=2.0.0,<3.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "pillow (>=12.1.0,<13.0.0)",
    "unidecode (>=1.4.0,<2.0.0)",
    "celery (>=5.6.2,<6.0.0)",
    "pandas (>=3.0.1,<4.0.0)",
    "redis (>=7.3.0,<8.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "django-redis (>=6.0.0,<7.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "pytest (>=9.0.2,<10.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.lint.dependencies]
black = "^26.1.0"
flake8 = "^7.3.0"
mypy = "^1.19.1"
isort = "^7.0.0"

[tool.black]
line-length = 119
exclude = '''
/(
    \.venv
  | venv
  | \.git
  | \.mypy_cache
  | \.pytest_cache
  | build
  | dist
)/
'''

[tool.isort]
line_length = 119

[tool.mypy]
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
exclude = 'venv|README\.md'
ignore_missing_imports = true

